+++ examples/C/zhelpers.h
@@ -99,7 +99,7 @@ s_dump (void *socket)
         }
         printf ("\n");
 
-        int64_t more;           //  Multipart detection
+        int more;        //  Multipart detection
         more = 0;
         size_t more_size = sizeof (more);
         zmq_getsockopt (socket, ZMQ_RCVMORE, &more, &more_size);
